requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.9
pytest==7.4.3
Werkzeug==3.0.1
gunicorn==21.2.0
boto3==1.35.19
minio==7.2.8
python-dotenv==1.0.1